removed cast

+++ OOPHomework/CommonTypeSystem/05.BitArray64/BitArray64.cs
@@ -165,7 +165,7 @@ namespace _05.BitArray64
             {
                 try
                 {
-                    return (int)values[position];
+                    return values[position];
                 }
                 catch (IndexOutOfRangeException)
                 {
